+++ CHANGELOG.md
@@ -10,6 +10,19 @@ Changed :
 
 - Default value for `GeoTooltip` delay
 
+## 26.1.4
+
+Fixed :
+
+- `deleteIcon` displayed when `GeoSelectToggleButton` is not empty
+
+## 26.1.3
+
+Fixed:
+
+- `GeoDropdown` will be properly right-aligned when its width changes.
+- `GeoInput` use 100% width.
+
 ## 26.1.2
 
 Fixed: